# contract/swagger.yaml
openapi: '3.0.0'
info:
  title: Bright Smiles Architects API
  version: '1.0.0'
  description: API for managing dental clinics
servers:
  - url: 'http://localhost:3001/api'
    description: Local development server
paths:
...
  /appointments:
    get:
      tags:
        - Appointments
      summary: Get appointments with search and filtering
      parameters:
        - name: query
          in: query
          required: false
          schema:
            type: string
          description: Full text search across appointment fields
        - name: status
          in: query
          required: false
          schema:
            $ref: '#/components/schemas/AppointmentStatus'
        - name: dateFrom
          in: query
          required: false
          schema:
            type: string
            format: date-time
        - name: dateTo
          in: query
          required: false
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AppointmentBrief'
    post:
      tags:
        - Appointments
      summary: Create a new appointment
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentCreateRequest'
      responses:
        '201':
          description: Appointment created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentDetails'
  /appointments/{appointmentId}:
    get:
      tags:
        - Appointments
      summary: Get appointment details
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Appointment details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AppointmentDetails'
        '404':
          description: Appointment not found
    put:
      tags:
        - Appointments
      summary: Update an appointment
      parameters:
        - name: appointmentId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AppointmentUpdateRequest'
      responses:
        '200':
          description: Appointment updated successfully
...
  /notifications/{notificationId}:
    get:
      tags:
        - Notifications
      summary: Get a single notification with full content
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Complete notification details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Notification'
  /notifications/{notificationId}/read:
    put:
      tags:
        - Notifications
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Notification marked as read successfully
...
components:
  schemas:
    BillingStatus:
      type: string
      enum: [PENDING, PAID, CANCELLED]
    AppointmentStatus:
      type: string
      enum: [SCHEDULED, COMPLETED, CANCELLED]
...
    Billing:
      type: object
      properties:
        amount:
          type: number
        status:
          $ref: '#/components/schemas/BillingStatus'
      required:
        - amount
        - status
    AppointmentDetails:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patient:
          $ref: '#/components/schemas/PatientBrief'
        doctor:
          $ref: '#/components/schemas/DoctorBrief'
        serviceType:
          $ref: '#/components/schemas/ServiceType'
        location:
          $ref: '#/components/schemas/ClinicBrief'
        datetime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
        notes:
          type: string
        prescriptions:
          type: array
          items:
            type: string
        billing:
          $ref: '#/components/schemas/Billing'
      required:
        - id
        - patient
        - doctor
        - serviceType
        - location
        - datetime
        - status
    AppointmentBrief:
      type: object
      properties:
        id:
          type: string
          format: uuid
        patientName:
          type: string
        doctorId:
          type: integer
        doctorName:
          type: string
        serviceType:
          type: string
        location:
          type: string
        datetime:
          type: string
          format: date-time
        status:
          $ref: '#/components/schemas/AppointmentStatus'
      required:
        - id
        - patient
        - doctorId
        - doctorName
        - serviceType
        - location
        - datetime
        - status
    AppointmentCreateRequest:
      type: object
      properties:
        patientId:
          type: string
          format: uuid
        doctorId:
          type: string
          format: uuid
        serviceTypeId:
          type: string
          format: uuid
        datetime:
          type: string
          format: date-time
      required:
        - patientId
        - doctorId
        - serviceTypeId
        - datetime
    AppointmentUpdateRequest:
      type: object
      required:
        - appointmentId
      properties:
        appointmentId:
          type: string
          format: uuid
        datetime:
          type: string
          format: date-time
        time:
          type: string
        description:
          type: string
        status:
          $ref: '#/components/schemas/AppointmentStatus'
...
    DoctorBrief:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the doctor
        fullName:
          type: string
          description: Full name of the doctor including titles (e.g. "Dr. John Smith")
        specialties:
          type: array
          description: List of doctor's areas of specialization (e.g. ["Orthodontics", "Dental Surgery"])
          items:
            type: string
        locations:
          type: array
          items:
            $ref: '#/components/schemas/ClinicBrief'
      required:
        - id
        - fullName
        - specialties
...
    PatientBrief:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique identifier for the patient
        fullName:
          type: string
          description: Full name of the patient
      required:
        - id
        - fullName
...
    ClinicBrief:
      type: object
      properties:
        id:
          type: integer
          description: Unique identifier for the location
        name:
          type: string
          description: Name of the clinic
        address:
          type: string
          description: Full address including country, city, street and postal code
      required:
        - id
        - name
        - address
...
    Notification:
      type: object
      properties:
        id:
          type: string
          format: uuid
        title:
          type: string
        subtitle:
          type: string
        receivedDate:
          type: string
          format: date
        read:
          type: boolean
          default: false
        content:
          type: string
      required:
        - id
        - title
        - subtitle
        - receivedDate
        - content
...
    ServiceType:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
      required:
        - id
        - name
...